requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
gevent==23.9.1
gevent-websocket==0.10.1
pytest==7.4.3
pytest-flask==1.3.0
